requirements.txt
# Core Dependencies
python-dotenv==1.2.1
pydantic==2.12.5
PyYAML==6.0.3
anyio>=4.4.0

# LLM Providers
openai==1.109.1
anthropic==0.76.0
requests==2.32.5

# Data Processing
numpy==1.26.4
pandas==2.0.0

# Database
sqlalchemy==2.0.45
psycopg2-binary==2.9.9

# API
httpx==0.28.1

# Logging
python-json-logger==2.0.7

# CLI
click==8.1.8
rich==13.9.4

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==24.3.0
flake8==6.1.0
mypy==1.7.1

sentence-transformers
tiktoken
